refactor: auth, common 스키마 글로벌 스키마 폴더로 이동 (#40)
