+++ src/hooks/queries/useAdminApplicationsQuery.ts
@@ -4,7 +4,7 @@ import {
   GetAdminApplicationsResponse,
 } from '@/schemas/admin/admin-applications-schema';
 import {getAdminApplications} from '@/services/api/admin/admin.application.api';
-import {ErrorResponse} from '@/services/schemas/common.schema';
+import {ErrorResponse} from '@/schemas/common/common-schema';
 import {useQuery} from '@tanstack/react-query';
 
 export const useAdminApplicationsQuery = (
